CAMEL-16064: preliminary refactoring of camel-kafka
This reorganizes streaming-related processing code to simplify the implementation of the batch processing code
