+++ components/camel-kafka/src/main/java/org/apache/camel/component/kafka/consumer/support/KafkaRecordProcessor.java
@@ -14,161 +14,14 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
 package org.apache.camel.component.kafka.consumer.support;
 
-import java.util.stream.StreamSupport;
-
-import org.apache.camel.Exchange;
-import org.apache.camel.Message;
-import org.apache.camel.Processor;
-import org.apache.camel.component.kafka.KafkaConfiguration;
-import org.apache.camel.component.kafka.KafkaConstants;
-import org.apache.camel.component.kafka.consumer.CommitManager;
-import org.apache.camel.component.kafka.consumer.KafkaManualCommit;
-import org.apache.camel.component.kafka.serde.KafkaHeaderDeserializer;
-import org.apache.camel.spi.ExceptionHandler;
-import org.apache.camel.spi.HeaderFilterStrategy;
-import org.apache.kafka.clients.consumer.ConsumerRecord;
+import org.apache.camel.component.kafka.KafkaConsumer;
 import org.apache.kafka.common.TopicPartition;
-import org.apache.kafka.common.header.Header;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-public class KafkaRecordProcessor {
-
-    private static final Logger LOG = LoggerFactory.getLogger(KafkaRecordProcessor.class);
-
-    private final boolean autoCommitEnabled;
-    private final KafkaConfiguration configuration;
-    private final Processor processor;
-    private final CommitManager commitManager;
-
-    public KafkaRecordProcessor(KafkaConfiguration configuration, Processor processor, CommitManager commitManager) {
-        this.autoCommitEnabled = configuration.isAutoCommitEnable();
-        this.configuration = configuration;
-        this.processor = processor;
-        this.commitManager = commitManager;
-    }
-
-    private void setupExchangeMessage(Message message, ConsumerRecord record) {
-        message.setHeader(KafkaConstants.PARTITION, record.partition());
-        message.setHeader(KafkaConstants.TOPIC, record.topic());
-        message.setHeader(KafkaConstants.OFFSET, record.offset());
-        message.setHeader(KafkaConstants.HEADERS, record.headers());
-        message.setHeader(KafkaConstants.TIMESTAMP, record.timestamp());
-        message.setHeader(Exchange.MESSAGE_TIMESTAMP, record.timestamp());
-
-        if (record.key() != null) {
-            message.setHeader(KafkaConstants.KEY, record.key());
-        }
-
-        LOG.debug("Setting up the exchange for message from partition {} and offset {}",
-                record.partition(), record.offset());
-
-        message.setBody(record.value());
-    }
-
-    private boolean shouldBeFiltered(Header header, Exchange exchange, HeaderFilterStrategy headerFilterStrategy) {
-        return !headerFilterStrategy.applyFilterToExternalHeaders(header.key(), header.value(), exchange);
-    }
-
-    private void propagateHeaders(ConsumerRecord<Object, Object> record, Exchange exchange) {
-
-        HeaderFilterStrategy headerFilterStrategy = configuration.getHeaderFilterStrategy();
-        KafkaHeaderDeserializer headerDeserializer = configuration.getHeaderDeserializer();
-
-        StreamSupport.stream(record.headers().spliterator(), false)
-                .filter(header -> shouldBeFiltered(header, exchange, headerFilterStrategy))
-                .forEach(header -> exchange.getIn().setHeader(header.key(),
-                        headerDeserializer.deserialize(header.key(), header.value())));
-    }
-
-    public ProcessingResult processExchange(
-            Exchange exchange, TopicPartition topicPartition, boolean partitionHasNext,
-            boolean recordHasNext, ConsumerRecord<Object, Object> record, ExceptionHandler exceptionHandler) {
-
-        Message message = exchange.getMessage();
-
-        setupExchangeMessage(message, record);
-
-        propagateHeaders(record, exchange);
-
-        // if not auto commit then we have additional information on the exchange
-        if (!autoCommitEnabled) {
-            message.setHeader(KafkaConstants.LAST_RECORD_BEFORE_COMMIT, !recordHasNext);
-            message.setHeader(KafkaConstants.LAST_POLL_RECORD, !recordHasNext && !partitionHasNext);
-        }
-
-        if (configuration.isAllowManualCommit()) {
-            // allow Camel users to access the Kafka consumer API to be able to do for example manual commits
-            KafkaManualCommit manual = commitManager.getManualCommit(exchange, topicPartition, record);
-
-            message.setHeader(KafkaConstants.MANUAL_COMMIT, manual);
-            message.setHeader(KafkaConstants.LAST_POLL_RECORD, !recordHasNext && !partitionHasNext);
-        }
-
-        try {
-            processor.process(exchange);
-        } catch (Exception e) {
-            exchange.setException(e);
-        }
-
-        ProcessingResult result = ProcessingResult.newUnprocessed();
-        if (exchange.getException() != null) {
-            LOG.debug("An exception was thrown for record at partition {} and offset {}",
-                    record.partition(), record.offset());
-
-            boolean breakOnErrorExit = processException(exchange, topicPartition, record, exceptionHandler);
-            result = new ProcessingResult(breakOnErrorExit, true);
-        } else {
-            result = new ProcessingResult(false, exchange.getException() != null);
-        }
-
-        if (!result.isBreakOnErrorHit()) {
-            commitManager.recordOffset(topicPartition, record.offset());
-        }
-
-        return result;
-    }
-
-    private boolean processException(
-            Exchange exchange, TopicPartition topicPartition,
-            ConsumerRecord<Object, Object> consumerRecord, ExceptionHandler exceptionHandler) {
-
-        // processing failed due to an unhandled exception, what should we do
-        if (configuration.isBreakOnFirstError()) {
-            // we are failing and we should break out
-            if (LOG.isWarnEnabled()) {
-                Exception exc = exchange.getException();
-                LOG.warn("Error during processing {} from topic: {} due to {}", exchange, topicPartition.topic(),
-                        exc.getMessage());
-                LOG.warn("Will seek consumer to offset {} on partition {} and start polling again.",
-                        consumerRecord.offset(), consumerRecord.partition());
-            }
-
-            // we should just do a commit (vs the original forceCommit)
-            // when route uses NOOP Commit Manager it will rely
-            // on the route implementation to explicitly commit offset
-            // when route uses Synch/Asynch Commit Manager it will
-            // ALWAYS commit the offset for the failing record
-            // and will ALWAYS retry it
-            commitManager.commit(topicPartition);
-
-            // continue to next partition
-            return true;
-        } else {
-            // will handle/log the exception and then continue to next
-            exceptionHandler.handleException("Error during processing", exchange, exchange.getException());
-        }
-
-        return false;
-    }
-
-    public static String serializeOffsetKey(TopicPartition topicPartition) {
-        return topicPartition.topic() + '/' + topicPartition.partition();
-    }
-
-    public static long deserializeOffsetValue(String offset) {
-        return Long.parseLong(offset);
-    }
+public interface KafkaRecordProcessor<T> {
+    ProcessingResult processExchange(
+            KafkaConsumer camelKafkaConsumer, TopicPartition topicPartition, boolean partitionHasNext,
+            boolean recordHasNext, T processable);
 }
