+++ components/camel-kafka/src/main/java/org/apache/camel/component/kafka/consumer/support/streaming/KafkaRecordStreamingProcessor.java
@@ -1,0 +1,175 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more
+ * contributor license agreements.  See the NOTICE file distributed with
+ * this work for additional information regarding copyright ownership.
+ * The ASF licenses this file to You under the Apache License, Version 2.0
+ * (the "License"); you may not use this file except in compliance with
+ * the License.  You may obtain a copy of the License at
+ *
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.camel.component.kafka.consumer.support.streaming;
+
+import java.util.stream.StreamSupport;
+
+import org.apache.camel.Exchange;
+import org.apache.camel.Message;
+import org.apache.camel.Processor;
+import org.apache.camel.component.kafka.KafkaConfiguration;
+import org.apache.camel.component.kafka.KafkaConstants;
+import org.apache.camel.component.kafka.KafkaConsumer;
+import org.apache.camel.component.kafka.consumer.CommitManager;
+import org.apache.camel.component.kafka.consumer.KafkaManualCommit;
+import org.apache.camel.component.kafka.consumer.support.KafkaRecordProcessor;
+import org.apache.camel.component.kafka.consumer.support.ProcessingResult;
+import org.apache.camel.component.kafka.serde.KafkaHeaderDeserializer;
+import org.apache.camel.spi.ExceptionHandler;
+import org.apache.camel.spi.HeaderFilterStrategy;
+import org.apache.kafka.clients.consumer.ConsumerRecord;
+import org.apache.kafka.common.TopicPartition;
+import org.apache.kafka.common.header.Header;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+
+final class KafkaRecordStreamingProcessor implements KafkaRecordProcessor<ConsumerRecord<Object, Object>> {
+    private static final Logger LOG = LoggerFactory.getLogger(KafkaRecordStreamingProcessor.class);
+
+    private final boolean autoCommitEnabled;
+    private final KafkaConfiguration configuration;
+    private final Processor processor;
+    private final CommitManager commitManager;
+
+    public KafkaRecordStreamingProcessor(KafkaConfiguration configuration, Processor processor, CommitManager commitManager) {
+        this.autoCommitEnabled = configuration.isAutoCommitEnable();
+        this.configuration = configuration;
+        this.processor = processor;
+        this.commitManager = commitManager;
+    }
+
+    private void setupExchangeMessage(Message message, ConsumerRecord<?, ?> consumerRecord) {
+        message.setHeader(KafkaConstants.PARTITION, consumerRecord.partition());
+        message.setHeader(KafkaConstants.TOPIC, consumerRecord.topic());
+        message.setHeader(KafkaConstants.OFFSET, consumerRecord.offset());
+        message.setHeader(KafkaConstants.HEADERS, consumerRecord.headers());
+        message.setHeader(KafkaConstants.TIMESTAMP, consumerRecord.timestamp());
+        message.setHeader(Exchange.MESSAGE_TIMESTAMP, consumerRecord.timestamp());
+
+        if (consumerRecord.key() != null) {
+            message.setHeader(KafkaConstants.KEY, consumerRecord.key());
+        }
+
+        LOG.debug("Setting up the exchange for message from partition {} and offset {}",
+                consumerRecord.partition(), consumerRecord.offset());
+
+        message.setBody(consumerRecord.value());
+    }
+
+    private boolean shouldBeFiltered(Header header, Exchange exchange, HeaderFilterStrategy headerFilterStrategy) {
+        return !headerFilterStrategy.applyFilterToExternalHeaders(header.key(), header.value(), exchange);
+    }
+
+    private void propagateHeaders(ConsumerRecord<Object, Object> consumerRecord, Exchange exchange) {
+
+        HeaderFilterStrategy headerFilterStrategy = configuration.getHeaderFilterStrategy();
+        KafkaHeaderDeserializer headerDeserializer = configuration.getHeaderDeserializer();
+
+        StreamSupport.stream(consumerRecord.headers().spliterator(), false)
+                .filter(header -> shouldBeFiltered(header, exchange, headerFilterStrategy))
+                .forEach(header -> exchange.getIn().setHeader(header.key(),
+                        headerDeserializer.deserialize(header.key(), header.value())));
+    }
+
+    @Override
+    public ProcessingResult processExchange(
+            KafkaConsumer camelKafkaConsumer, TopicPartition topicPartition, boolean partitionHasNext,
+            boolean recordHasNext, ConsumerRecord<Object, Object> consumerRecord) {
+
+        final Exchange exchange = camelKafkaConsumer.createExchange(false);
+        final ExceptionHandler exceptionHandler = camelKafkaConsumer.getExceptionHandler();
+
+        Message message = exchange.getMessage();
+
+        setupExchangeMessage(message, consumerRecord);
+
+        propagateHeaders(consumerRecord, exchange);
+
+        // if not auto commit then we have additional information on the exchange
+        if (!autoCommitEnabled) {
+            message.setHeader(KafkaConstants.LAST_RECORD_BEFORE_COMMIT, !recordHasNext);
+            message.setHeader(KafkaConstants.LAST_POLL_RECORD, !recordHasNext && !partitionHasNext);
+        }
+
+        if (configuration.isAllowManualCommit()) {
+            // allow Camel users to access the Kafka consumer API to be able to do for example manual commits
+            KafkaManualCommit manual = commitManager.getManualCommit(exchange, topicPartition, consumerRecord);
+
+            message.setHeader(KafkaConstants.MANUAL_COMMIT, manual);
+            message.setHeader(KafkaConstants.LAST_POLL_RECORD, !recordHasNext && !partitionHasNext);
+        }
+
+        try {
+            processor.process(exchange);
+        } catch (Exception e) {
+            exchange.setException(e);
+        }
+
+        ProcessingResult result;
+        if (exchange.getException() != null) {
+            LOG.debug("An exception was thrown for consumerRecord at partition {} and offset {}",
+                    consumerRecord.partition(), consumerRecord.offset());
+
+            boolean breakOnErrorExit = processException(exchange, topicPartition, consumerRecord, exceptionHandler);
+            result = new ProcessingResult(breakOnErrorExit, true);
+        } else {
+            result = new ProcessingResult(false, exchange.getException() != null);
+        }
+
+        if (!result.isBreakOnErrorHit()) {
+            commitManager.recordOffset(topicPartition, consumerRecord.offset());
+        }
+
+        // success so release the exchange
+        camelKafkaConsumer.releaseExchange(exchange, false);
+
+        return result;
+    }
+
+    private boolean processException(
+            Exchange exchange, TopicPartition topicPartition,
+            ConsumerRecord<Object, Object> consumerRecord, ExceptionHandler exceptionHandler) {
+
+        // processing failed due to an unhandled exception, what should we do
+        if (configuration.isBreakOnFirstError()) {
+            // we are failing and we should break out
+            if (LOG.isWarnEnabled()) {
+                Exception exc = exchange.getException();
+                LOG.warn("Error during processing {} from topic: {} due to {}", exchange, topicPartition.topic(),
+                        exc.getMessage());
+                LOG.warn("Will seek consumer to offset {} on partition {} and start polling again.",
+                        consumerRecord.offset(), consumerRecord.partition());
+            }
+
+            // we should just do a commit (vs the original forceCommit)
+            // when route uses NOOP Commit Manager it will rely
+            // on the route implementation to explicitly commit offset
+            // when route uses Synch/Asynch Commit Manager it will
+            // ALWAYS commit the offset for the failing record
+            // and will ALWAYS retry it
+            commitManager.commit(topicPartition);
+
+            // continue to next partition
+            return true;
+        } else {
+            // will handle/log the exception and then continue to next
+            exceptionHandler.handleException("Error during processing", exchange, exchange.getException());
+        }
+
+        return false;
+    }
+}
